Changes from review comments, moved guidance_eval_save to base.py
