+++ threestudio/models/guidance/stable_diffusion_guidance.py
@@ -431,9 +431,9 @@ class StableDiffusionGuidance(BaseObject):
 
         if guidance_eval:
             guidance_eval_out = self.guidance_eval(**guidance_eval_utils)
-            return guidance_out, guidance_eval_out
-        else:
-            return guidance_out
+            guidance_out.update({"eval": guidance_eval_out})
+
+        return guidance_out
 
     @torch.cuda.amp.autocast(enabled=False)
     @torch.no_grad()
