+++ threestudio/systems/base.py
@@ -2,6 +2,7 @@ import os
 from dataclasses import dataclass, field
 
 import pytorch_lightning as pl
+import torch.nn.functional as F
 
 import threestudio
 from threestudio.models.exporters.base import Exporter, ExporterOutput
@@ -305,3 +306,63 @@ class BaseLift3DSystem(BaseSystem):
     def on_predict_end(self) -> None:
         if self._save_dir is not None:
             threestudio.info(f"Export assets saved to {self._save_dir}")
+
+    def guidance_evaluation_save(self, comp_rgb, guidance_eval_out):
+        B, size = comp_rgb.shape[:2]
+        resize = lambda x: F.interpolate(
+            x.permute(0, 3, 1, 2), (size, size), mode="bilinear", align_corners=False
+        ).permute(0, 2, 3, 1)
+        filename = f"it{self.true_global_step}-train.png"
+
+        def merge12(x):
+            return x.reshape(-1, *x.shape[2:])
+
+        self.save_image_grid(
+            filename,
+            [
+                {
+                    "type": "rgb",
+                    "img": merge12(comp_rgb),
+                    "kwargs": {"data_format": "HWC"},
+                },
+            ]
+            + (
+                [
+                    {
+                        "type": "rgb",
+                        "img": merge12(resize(guidance_eval_out["imgs_noisy"])),
+                        "kwargs": {"data_format": "HWC"},
+                    }
+                ]
+            )
+            + (
+                [
+                    {
+                        "type": "rgb",
+                        "img": merge12(resize(guidance_eval_out["imgs_1step"])),
+                        "kwargs": {"data_format": "HWC"},
+                    }
+                ]
+            )
+            + (
+                [
+                    {
+                        "type": "rgb",
+                        "img": merge12(resize(guidance_eval_out["imgs_1orig"])),
+                        "kwargs": {"data_format": "HWC"},
+                    }
+                ]
+            )
+            + (
+                [
+                    {
+                        "type": "rgb",
+                        "img": merge12(resize(guidance_eval_out["imgs_final"])),
+                        "kwargs": {"data_format": "HWC"},
+                    }
+                ]
+            ),
+            name="train_step",
+            step=self.true_global_step,
+            noise_levels=guidance_eval_out["noise_levels"],
+        )
